+++ probdiffeq/probdiffeq.py
@@ -4,7 +4,10 @@ See the tutorials for example use cases.
 """
 
 from probdiffeq import diffeqjet, ssm_impl
-from probdiffeq.backend import flow, func, inspect, linalg, np, random, structs, tree
+from probdiffeq._probdiffeq.jacobian_handlers import *
+from probdiffeq._probdiffeq.loss_functions import *
+from probdiffeq._probdiffeq.markov_processes import *
+from probdiffeq.backend import func, inspect, linalg, np, structs, tree
 from probdiffeq.backend.typing import (
     Any,
     Array,
@@ -27,261 +30,6 @@ N = TypeVar("N", bound=ssm_impl.AbstractTreeNormal)
 
 Used to type marginals, for example.
 """
-
-
-class JacobianHandler:
-    """An interface for working with Jacobian matrices."""
-
-    def init_jacobian_handler(self):
-        """Initialize the handler state.
-
-        For example, if the handler uses stochastic sampling,
-        this initialisation would create a random key.
-        """
-        raise NotImplementedError
-
-    def materialize_dense(self, fun, x, state, /):
-        """Materialize a dense Jacobian.
-
-        This is typically used for first-order linearization in dense
-        state-space models.
-        """
-        raise NotImplementedError
-
-    def calculate_trace(self, fun, x, state, /):
-        """Calculate the trace of a Jacobian.
-
-        This is typically used for first-order linearization in isotropic
-        state-space models.
-        """
-        raise NotImplementedError
-
-    def calculate_diagonal(self, fun, x, state, /):
-        """Calculate the diagonal of a Jacobian.
-
-        This is typically used for first-order linearization in block-diagonal
-        state-space models.
-        """
-        raise NotImplementedError
-
-
-class jacobian_materialize(JacobianHandler):
-    """Construct a handler that always materialized Jacobian matrices.
-
-    Use this Jacobian if the dimension of the problem is relatively small.
-    """
-
-    def __init__(self, *, jacfun=func.jacfwd) -> None:
-        self.jacfun = jacfun
-
-    def init_jacobian_handler(self):
-        return ()
-
-    def materialize_dense(self, fun, x, state, /):
-        del state
-        fx = fun(x)
-        dfx = func.jacfwd(fun)(x)
-        return fx, dfx, ()
-
-    def calculate_trace(self, fun, x, state, /):
-        del state
-        fx = fun(x)
-        dfx = func.jacfwd(fun)(x)
-        dfx_trace = linalg.trace(dfx)
-        return fx, dfx_trace, ()
-
-    def calculate_diagonal(self, fun, x, state, /):
-        del state
-        fx = fun(x)
-        dfx = func.jacfwd(fun)(x)
-        dfx_diagonal = linalg.diagonal(dfx)
-        return fx, dfx_diagonal, ()
-
-
-class jacobian_hutchinson_fwd(JacobianHandler):
-    """Construct a handler that uses stochastic trace estimation for traces/diagonals.
-
-    Use a Hutchinson handler if the dimension of the problem is large.
-
-    This implementation uses **forward-mode** automatic differentiation.
-
-    !!! warning "Warning: highly EXPERIMENTAL feature!"
-        This function is highly experimental and not safe to use.
-        There is no guarantee that it works correctly (or at all).
-        It might be deleted tomorrow and without any deprecation policy.
-
-    """
-
-    def __init__(self, *, seed=1, num_probes=10) -> None:
-        self.seed = seed
-        self.num_probes = num_probes
-
-    def init_jacobian_handler(self):
-        return random.prng_key(seed=self.seed)
-
-    def materialize_dense(self, fun, x, state, /):
-        # TODO: approximate Jacobian with outer products instead of forming?
-        # What is the "correct" thing to do?
-        fx = fun(x)
-        dfx = func.jacfwd(fun)(x)
-        return fx, dfx, state
-
-    def calculate_trace(self, fun, x, key, /):
-        key, subkey = random.split(key, num=2)
-        sample_shape = (self.num_probes, *x.shape)
-        v = random.rademacher(subkey, shape=sample_shape, dtype=x.dtype)
-
-        fx, Jvp = func.linearize(fun, x)
-        J_trace = func.vmap(lambda s: linalg.vector_dot(s, Jvp(s)))(v)
-        J_trace = J_trace.mean(axis=0)
-        return fx, J_trace, key
-
-    def calculate_diagonal(self, fun, x, key, /):
-        key, subkey = random.split(key, num=2)
-        sample_shape = (self.num_probes, *x.shape)
-        v = random.rademacher(subkey, shape=sample_shape, dtype=x.dtype)
-
-        fx, Jvp = func.linearize(fun, x)
-        vJv = func.vmap(lambda s: s * Jvp(s))(v)
-        J_diagonal = vJv.mean(axis=0)
-        return fx, J_diagonal, key
-
-
-class jacobian_hutchinson_rev(JacobianHandler):
-    """Construct a handler that uses stochastic trace estimation for traces/diagonals.
-
-    Use a Hutchinson handler if the dimension of the problem is large.
-
-    This implementation uses **reverse-mode** automatic differentiation.
-
-    !!! warning "Warning: highly EXPERIMENTAL feature!"
-        This function is highly experimental and not safe to use.
-        There is no guarantee that it works correctly (or at all).
-        It might be deleted tomorrow and without any deprecation policy.
-
-    """
-
-    def __init__(self, *, seed=1, num_probes=10) -> None:
-        self.seed = seed
-        self.num_probes = num_probes
-
-    def init_jacobian_handler(self):
-        return random.prng_key(seed=self.seed)
-
-    def materialize_dense(self, fun, x, state, /):
-        # TODO: approximate Jacobian with outer products instead of forming?
-        # What is the "correct" thing to do?
-        fx = fun(x)
-        dfx = func.jacrev(fun)(x)
-        return fx, dfx, state
-
-    def calculate_trace(self, fun, x, key, /):
-        key, subkey = random.split(key, num=2)
-        sample_shape = (self.num_probes, *x.shape)
-        v = random.rademacher(subkey, shape=sample_shape, dtype=x.dtype)
-
-        fx, vjp = func.vjp(fun, x)
-        J_trace = func.vmap(lambda s: linalg.vector_dot(s, vjp(s)[0]))(v)
-        J_trace = J_trace.mean(axis=0)
-        return fx, J_trace, key
-
-    def calculate_diagonal(self, fun, x, key, /):
-        key, subkey = random.split(key, num=2)
-        sample_shape = (self.num_probes, *x.shape)
-        v = random.rademacher(subkey, shape=sample_shape, dtype=x.dtype)
-
-        fx, vjp = func.vjp(fun, x)
-        vJv = func.vmap(lambda s: s * vjp(s)[0])(v)
-        J_diagonal = vJv.mean(axis=0)
-        return fx, J_diagonal, key
-
-
-def loss_lml_terminal_values(*, ssm: ssm_impl.FactSsmImpl, tcoeff_index=0):
-    """Construct a log-marginal-likelihood loss for the terminal value."""
-
-    def loss(u, /, *, marginals, std):
-        u = tree.tree_map(np.asarray, u)
-        std = tree.tree_map(np.asarray, std)
-        std_expected = marginals.std[tcoeff_index]
-
-        msg = "The standard deviation container differs from what was expected."
-        msg += f" Expected: shape={tree.tree_map(np.shape, std_expected)}."
-        msg += f" Received: shape={tree.tree_map(np.shape, std)}."
-        msg += f" For reference: data-shape={tree.tree_map(np.shape, u)}."
-
-        try:
-            shapes_equal = tree.tree_map(
-                lambda a, b: a.shape == b.shape, std, std_expected
-            )
-        except Exception as error:
-            raise ValueError(msg) from error
-
-        if not tree.tree_all(shapes_equal):
-            raise ValueError(msg)
-
-        model = ssm.prior.to_derivative(tcoeff_index, std, template=marginals)
-        marg = ssm.conditional.marginalise(marginals, model)
-
-        # Wrap into list because blockdiag & isotropic models
-        # expect sequences of states (even if the length is one)
-        return marg.logpdf_tree([u])
-
-    return loss
-
-
-def loss_lml_timeseries(
-    *,
-    ssm: ssm_impl.FactSsmImpl,
-    average_pdfs: bool = True,
-    tcoeff_index=0,
-    solve_triu=linalg.lstsq_svd,
-):
-    """Construct a log-marginal-likelihood loss for a time-series."""
-
-    def loss(u, /, *, posterior, std):
-        if not isinstance(posterior, MarkovSequence):
-            msg = "The datatype of the posterior is not as expected."
-            msg += f" Expected: {MarkovSequence}."
-            msg += f" Received: {type(posterior)}."
-            msg += " Did you perhaps use a filter instead of a smoother"
-            msg += ", or mean to use a different loss?"
-            raise TypeError(msg)
-
-        u = tree.tree_map(np.asarray, u)
-        std = tree.tree_map(np.asarray, std)
-        std_expected = posterior.marginal.std[tcoeff_index]
-
-        msg = "The standard deviation container differs from what was expected."
-        msg += f" Expected: shape={tree.tree_map(np.shape, std_expected)}."
-        msg += f" Received: shape={tree.tree_map(np.shape, std)}."
-        msg += f" For reference: data-shape={tree.tree_map(np.shape, u)}."
-
-        try:
-            shapes_equal = tree.tree_map(
-                lambda a, b: a.shape == b.shape, std, std_expected
-            )
-        except Exception as error:
-            raise ValueError(msg) from error
-
-        if not tree.tree_all(shapes_equal):
-            raise ValueError(msg)
-
-        # Remove the filtering distributions from the posterior
-        posterior = posterior.remove_filtering_distributions()
-
-        def make_model(s):
-            return ssm.prior.to_derivative(tcoeff_index, s, template=posterior.marginal)
-
-        model = func.vmap(make_model)(std)
-
-        # Use solve_triu=lstsq because for noise-free observations, the initial state
-        # of the ODE solution tends to be noise-free,
-        # which clashes and returns NaNs if we use exact solvers.
-        return posterior.evaluate_lml(
-            [u], model=model, ssm=ssm, average_pdfs=average_pdfs, solve_triu=solve_triu
-        )
-
-    return loss
 
 
 class Constraint(Protocol):
@@ -656,163 +404,6 @@ def constraint_jet_dae(
     return ssm.linearize.root(
         root_jet, root_order=order, jacobian=jacobian, nlstsq=nlstsq
     )
-
-
-@tree.register_dataclass
-@structs.dataclass
-class MarkovSequence(Generic[N]):
-    """A datastructure for Markov sequences as batches of joint distributions.
-
-    This is the output type of smoother-based estimators.
-    (Filter-based estimators do not return specialised types.)
-    """
-
-    marginal: N
-    """The marginal distribution."""
-
-    conditional: Any
-    """The conditional distribution."""
-
-    reverse: bool = structs.dataclass_field(metadata={"static": True})
-    """The direction of factorisations."""
-
-    @classmethod
-    def from_grid(cls, init, discretize, *, grid, reverse: bool):
-        marginal = init
-        conditional = func.vmap(discretize)(np.diff(grid))
-        return cls(marginal, conditional, reverse=reverse)
-
-    def rescale_cholesky(self, factor, /):
-        marg = self.marginal.rescale_cholesky(factor)
-        cond = self.conditional.rescale_noise(factor)
-        return MarkovSequence(marg, cond, reverse=self.reverse)
-
-    def evaluate_marginals(self, *, ssm):
-        """Extract the (time-)marginals from a Markov sequence.
-
-        This is only needed in combination with smoothing-based strategies.
-        """
-        if self.marginal.mean_flat.ndim == self.conditional.noise.mean_flat.ndim:
-            # TODO: do this in the postprocessing of the solver...
-            markov_seq = self.remove_filtering_distributions()
-            return markov_seq.evaluate_marginals(ssm=ssm)
-
-        def step(x, cond):
-            extrapolated = ssm.conditional.marginalise(x, cond)
-            return extrapolated, extrapolated
-
-        _, marginals = flow.scan(
-            step, init=self.marginal, xs=self.conditional, reverse=self.reverse
-        )
-
-        if self.reverse:
-            # Append the terminal marginal to the computed ones
-            return tree.tree_array_append(marginals, self.marginal)
-
-        return tree.tree_array_prepend(self.marginal, marginals)
-
-    def evaluate_lml(
-        self,
-        u,
-        *,
-        model,
-        ssm: ssm_impl.FactSsmImpl,
-        average_pdfs: bool,
-        solve_triu: Callable,
-    ):
-        assert self.reverse
-
-        # Process the terminal value
-        u0 = tree.tree_map(lambda s: s[-1], u)
-        model0 = tree.tree_map(lambda s: s[-1], model)
-        pdf0, updated = ssm.conditional.bayes_rule_and_logpdf_tree(
-            u0, self.marginal, model0, solve_triu=solve_triu
-        )
-
-        # Process the remaining values
-        def body(rv_and_logpdf, prior_and_observation_and_data):
-            rv, logpdf, num_data = rv_and_logpdf
-            prior, observation, data = prior_and_observation_and_data
-
-            predicted = ssm.conditional.marginalise(rv, prior)
-
-            logpdf_n, corrected = ssm.conditional.bayes_rule_and_logpdf_tree(
-                data, predicted, observation, solve_triu=solve_triu
-            )
-
-            # The mean of the PDFs (as opposed to their sum) usually
-            # leads to LML values that are more "human-readable"
-            # (ie magnitude O(1) instead O(N)). This is technically not
-            # a log-marginal-likelihood, but much nicer to work with when
-            # optimising the loss.
-            if average_pdfs:
-                logpdf1 = (logpdf * num_data + logpdf_n) / (num_data + 1)
-            else:
-                logpdf1 = logpdf + logpdf_n
-
-            return (corrected, logpdf1, num_data + 1), ()
-
-        u1 = tree.tree_map(lambda s: s[:-1], u)
-        model1 = tree.tree_map(lambda s: s[:-1], model)
-        init = (updated, pdf0, 1)
-        xs = (self.conditional, model1, u1)
-        (_, pdf, _), _ = flow.scan(body, init=init, xs=xs, reverse=self.reverse)
-        return pdf
-
-    def remove_filtering_distributions(self):
-        """Discard all intermediate filtering solutions from a Markov sequence.
-
-        This function is useful to convert a smoothing-solution into a Markov sequence
-        that is compatible with sampling or marginalisation.
-        """
-        # There should be one marginal and many conditionals. If not, remove marginals
-        if self.marginal.mean_flat.ndim == self.conditional.noise.mean_flat.ndim:
-            marginal_idx = -1 if self.reverse else 0
-            init = tree.tree_map(lambda x: x[marginal_idx, ...], self.marginal)
-            return MarkovSequence(init, self.conditional, reverse=self.reverse)
-        return self
-
-    def sample(self, key, *, ssm: ssm_impl.FactSsmImpl, shape: tuple = ()):
-        """Sample from a Markov sequence."""
-        # If the MarkovSequence carries unnecessary filtering marginals, remove them
-        if self.marginal.mean_flat.ndim == self.conditional.noise.mean_flat.ndim:
-            markov_seq = self.remove_filtering_distributions()
-            return markov_seq.sample(key, ssm=ssm, shape=shape)
-
-        # If many samples are required, vmap over recursive calls to sample()
-        if len(shape) > 0:
-            n, *shape_remaining = shape
-            keys = random.split(key, num=n)
-            sample_partial = func.partial(self.sample, ssm=ssm, shape=shape_remaining)
-            return func.vmap(sample_partial)(keys)
-
-        # Compute a single sample from the Markov sequence
-
-        def body(sample_and_key, cond):
-            smp_flat, key = sample_and_key
-
-            # Propagate the previous sample
-            predicted = ssm.conditional.apply_flat(smp_flat, cond)
-
-            # Sample the propagated variable
-            key, subkey = random.split(key, num=2)
-            smp_flat = predicted.sample_flat(subkey)
-
-            # Unravel the sample
-            smp_tree = predicted.tree_flatten.unflatten_array(smp_flat)
-            return (smp_flat, key), smp_tree
-
-        # Loop over the conditionals
-        key, subkey = random.split(key, num=2)
-        sample0_flat = self.marginal.sample_flat(subkey)
-        init = (sample0_flat, key)
-        xs = self.conditional
-        _, samples = flow.scan(body, init=init, xs=xs, reverse=self.reverse)
-
-        sample0_tree = self.marginal.tree_flatten.unflatten_array(sample0_flat)
-        if self.reverse:
-            return tree.tree_array_append(samples, sample0_tree)
-        return tree.tree_array_prepend(sample0_tree, samples)
 
 
 T = TypeVar("T", bound=MarkovSequence | ssm_impl.AbstractTreeNormal)
